Track idle_timeout events in Cowboy
When a request is binned by Cowboy because `idle_timeout` is reached, a
telemetry event `cowboy.request.stop` is emitted with an `error:
{:connection_error, :timeout, _human_friendly_message}` element in the
metadata is emitted. We'll now listen to that event and increment the
`cowboy.request.idle_timeout.count` metric. This will allow us to see
how often this happens.