Fix the way tags are set for smoke tests
We set `route` and `stack` tags in smoke tests to be able to filter them
accordingly. Smoke tests are filtered when they are run on a deployment:
tests are run against the deployed stack/env in that case.

ExUnit tags work roughly like this:

* Tags are module attributes that can
[accumulate](https://hexdocs.pm/elixir/Module.html#register_attribute/3)
values.
* When a `tag` attribute is set in a test file, the tag value is stored
in the test module. Multiple tags can be set because values are
accumulated.
* `test` is a macro and when it's called, the `tag` module attribute
value is read and stored by ExUnit with the contents of the test. The
`tag` attribute is then cleared to ensure that each test has its own
tags.

In short, tags are temporary storage of metadata for tests in a test
module. Values are accummuated with every `@tag` line in a test file and
cleared with every call to the `test` macro.

Smoke tests for each route matcher are defined in a loop: one test for
each route example. We need to define tags once for each invocation of
the `test` macro, i.e. inside the loop, otherwise they'll only be used
by the first test (then cleared and the rest of the tests in the loop
won't have the tags).

Before this change this meant than the tags would only be defined for
the first route example and so only that example would be used when
smoke test were filtered using tags.