Don't redirect WS images to .com
These static image URLs are redirected in Mozart to bbci.co.uk so
there's no need to redirect them to .com prior to sending the request to
Mozart.

I'm making this change to fix the smoke test for this route which
expects the redirect to .com but doesn't set the request host to
`.co.uk` for the redirect to happen. I thought that instead of fixing
this it'd be right to just disable the redirect for this route.