Set x-forwarded-host to *.com in smoke tests for WorldService routes
Before this change we'd set the `x-forwarded-host` header when
requesting WorldService routes in smoke tests to a `.co.uk` domain and
that would trigger a redirect in Belfrage to the corresponding `.com`
domain. We'd then verify that the redirect occured.

This change ensures that we set the `x-forwarded-host` header to a
`.com` domain, which means that the request is not redirected by
Belfrage, but passed through to the origin, like it's done for all other
routes.

I also removed the code that verified `.com` to `.co.uk` redirects
because that never happens in smoke tests: we never set the host to a
`.com` domain for routes that are supposed to be redirected to `.co.uk`.

Finally I moved the code that remained in `Support.Smoke.Assertions` to
`Belfrage.SmokeTestCase`: there wasn't much code left, so having it in a
separate module didn't make sense.