Track metrics of MachineGun pools individually
We previously collected metrics of `MachineGun` pools of http workers
(that are implemented using `:poolboy`) in `Metrics.PoolObserver` which
would run periodically and track the minimum number of available workers
and the maximum number of overflow workers across all pools. That
effectively tracked the busiest pool and it wasn't clear which exactly
pool that was and how the other pools were doing.

We'll now track `MachineGun`'s `:poolboy` pools individually using
`:telemetry_poller` which will collect the metrics periodically as
before, but emit individual measurements for each pool which will then
be tracked as separate metrics.

The new metric names are:

* `poolboy.available_workers.count.X`
* `poolboy.overflow_workers.count.X`

where `X` is the name of the endpoint that the pool is for. Endpoints
can be origins or other services that Belfrage uses (e.g. the auth
flagpole).

The name of the endpoint is determined from the name of the `:poolboy`
process, which is set by `MachineGun` and starts with the name of the
configuration. Because of this I've added separate `MachineGun`
configurations for each endpoint we're using. I thought that this would
be the most straightforward way to identify the pools, but also has an
advantage of fine-tuning the configuration for endpoints (e.g. we
probably don't need the same number of workers for all endpoints).

The name of the endpoint is appended to the metric name by
`:telemetry_metrics` when it groups the metrics by the configured tag
(the name of the pool). This is a bit different to some other metrics we
have and is not configurable, but I thought that the fact that we get
the metrics-related functionality out-of-the-box outweighs this.

Also the metric name starts with `poolboy` because we're planning to
track other `:poolboy` pools as well, not just `MachineGun` pools that
re used for making http requests.

I made the new poller track metrics every 5 seconds instead of 10s,
because I think we probably want to do that more frequently.

Finally, I added `Belfrage.Metrics` module that encapsulates the
conventions around using `:telemetry` in Belfrage (we'll keep extending
it to provide more functionality) and a helper module for testing
events.