Don't dynamically generate tests in Routes.RoutefileTest
Before this change we'd dynamically define multiple tests for every
route and redirect in the routefile, which was very slow because
`ExUnit.Case.test` is a macro that is expanded at complitaion time and
`ExUnit` compiles test files when loading them, so passing it the test
code multiple times for every route/redirect resulted in that test code
repeated hundreds of times in the resulting module during compilation,
i.e we ended up with a huge module to compile. In short the slowness
occurred during compilation, not during execution.

Instead of defining tests for every route/redirect we can have tests
that iterate through them and perform checks. Defining individual tests
seems like a good idea because it provides clear output on failure which
shows which route failed and why because there is a single test for each
route/redirect. However, we can achieve similar output with some custom
helpers that accumulate and format errors instead. I reckon the
difference in performance (these tests would previously take ~3 minutes
to run and now complete within 1.5s) is worth it.

Another reason to not define tests dynamically is that we also had to
hack our test config in `mix.exs` and use a different extension (`.ex`)
for routes tests (and smoke tests which use the same approach) to ensure
that they are not loaded with other tests (because even if we filtered
them, they'd still need to be compiled and that's what was realy slow).
We can update that configuration now.