Add more detailed tracking of internal latency
This commit tracks duration of various operations that happen during an
event lifecycle and reports them to CloudWatch. The idea is to record
these timings to see how they change over time or during incidents.

The operations are what I thought were logical, independant bits of
non-trivial logic, so this is just one way of splitting processing of a
request in Belfrage into parts. We might want to tweak these, remove or
add new operations as we gather some data / knowledge.

These operations are not a breakdown of latency of an average request
because not all requests go through all operations (e.g. some are
cached, compressed/decompressed, etc and others aren't), but these
metrics should allow us to see how expensive various operations are
(relative to each other), compare them and identify candidates for
optimisation. There's also some overlap in these operations (e.g.
`request_pipeline` includes `parse_session_token` because I thought that
the latter was important enough to track, but not other request
transformers individually).

Currently the following potentially important bits are not included
because there is no straightforward way of tracking them:

* `:cowboy` latency (i.e. anything that happens before and after a
request is passed to Belfrage)
* Internal latency of making a request to an origin (this requires more
accurate tracking of requests to origins)

The metrics are reported in microseconds because many of these
operations seem to take less than a millisecond.