Fix Benchmark.ResponsePipeline compilation warning
The following warning is displayed when `mix compile` is run:

    Compiling 5 files (.ex)
    warning: Belfrage.Services.Webcore.Response.build/1 is undefined or private. Did you mean one of:

          * build/2

    Found at 2 locations:
      benchmark/response_pipeline.ex:52: Benchmark.ResponsePipeline.setup/2
      benchmark/response_pipeline.ex:53: Benchmark.ResponsePipeline.setup/2

The purpose of this benchmark is to compare non-gzip and gzip
(base64 encoded) origin responses. A tuple with the shape:
`{:ok, _}` and not `{:error, _}` is passed to `build/2` in both
cases, so the value of the second argument does not matter.