Handle invalid URI encoded query string when checking it's valid UTF8 (#1037)
We check that request path and query string are valid UTF8 strings and
log them otherwise. The code that performs the check has to URI decode
them first and we need to handle incorrectly encoded URI strings there
as otherwise that code raises an exception.

This commit swallows the exception that is raised: it looks like there's
nothing else we can do. We can't tell if the string contains invalid
UTF8 characters if we can't URI decode it and `Plug` raises a
`Plug.Conn.InvalidQueryError` further down the line which we handle and
log (confirmed by the test).

Co-authored-by: David Lewis <87661013+thymusvulgaris@users.noreply.github.com>