[RESFRAME-4404] Populate in-memory cache when fetching fallbacks
If a request is successfully made to an origin (status code: 200)
then the response will be cached in the local and distributed
caches. If the same request is made again and:

* the request fails with a status code of 500
* the cached response for the request has expired in the
  local cache

then the request will be fetched from the distributed cache.
Subsequent requests that fail will also be fetched from the
distributed cache. In order to reduce repeated calls for the same
cached response in the distributed cache, this commit stores
the response in the local cache when a response has been fetched
from the distributed cache.

This commit achieves this by:

* Including the cache type with the response fetched from
  a cache - this can either be :local or :distributed.

* Setting the new `struct.response.cache_type` field to the
  cache type when a stale response is fetched.

* Storing the response in the local cache if the
  `struct.response.fallback` is `true` and the
  `struct.response.cache_type` is `:distributed`.

To prevent an early response from being fetched, the response
stored in the local cache has to be marked as stale.

Tests for `Belfrage.Cache.Fetch.fetch/2` are also added.