Don't track latency of fetching fallbacks
I added it along with all the other operations in #1036, but the way I
did it meant that it included the timing of requests to S3, so the value
of this metric is way higher than others (and so it didn't make sense to
include it on the graphs because it'd always dominate) and we also
actually already track this as `service.S3.request.timing`.

It would be better to add tracking of internal latency that occurs when
we fetch a fallback from S3 at some point in future if possible.