Add personalisation smoke tests (#1054)
* These new smoke tests will test 3 things on both test and live:

* A user with an invalid session is redirected to sign-in page.
* A signed-in user requesting the homepage gets a personalised response.
* A non-signed-in user requesting the homepage gets a non-personalised
response.

To determine if a response is personalised or not, the tests check the
`cache-control` response header: if it's `private` then the response is
personalised. A more sophisticated way to test that a response is
personalised would be to actually parse the content and see if any of
the personalisation features are present on the homepage, however that
is quite complex and is subject to change, so I've decided to go with
the simple approach for now.

Tests get an access token by emulating authentication by a user with
email/password that are used for testing personalisation on test & live.
This approach has been recommended by the Account team.

Tests use `:httpc` to make HTTP requests and not `MachineGun` like other
smoke tests because:

* I had a problem with setting request headers with `MachineGun` and it
doesn't provide any debugging capability.
* I thought that it doesn't make sense to maintain a pool of connections
in smoke tests which `MachineGun` does.