Refactor implementation of cascade (#1066)
* Consolidate all casade logic in Belfrage.Cascade

I thought that there were these issues with the implementation of the
cascade functionality:

* The logic was present in 4 different places: `Belfrage`,
`Belfrage.Processor`, `Belfrage.Concurrently` and
`Belfrage.Services.Cascade`. Having it spread across those modules made
it harder to understand the scope of the functionality and how it
worked.
* Naming of some of the modules/functions was confusing: e.g. only one
of the modules involved (`Beflrage.Services.Cascade`) had the word
"cascade" in it, which could lead you to believe that it's the only
module relevant to the cascade. Also, the name `Belfrage.Concurrently`
sounded like it was a general-purpose helper for concurrency, when in
fact it was part of the implementation of the cascade and could not be
used for anything else.

This change addresses the issues by consolidating the logic in a single
module called `Belfrage.Cascade` which will now be used in `Belfrage`
module only. It also updates the naming to be more consistent with the
rest of the codebase and hopefully clearer.

While making this change I also:

* Added `Belfrage.ServiceProvider.dispatch/1` function to make it easier 
to use dependency injection when testing `Befrage.Cascade.dispatch/1`
(otherwise it would need to mock both `ServiceProvider` and individual
services).
* Removed `Behaviours.ServiceProvider`. There was only one module 
that implemented this behaviour and it's not mocked so I'd say the 
behaviour is redundant.