Add Cascade struct (#1069)
* Add Cascade struct

When discussing the introduction of the Cascade module with the team,
the following became apparent to me:

* Cascade functionality is not a temporary feature: it will stay and
will continue to help capability teams to migrate to Webcore. Therefore
the ability to route a request to multiple origins should be treated as
an integral feature of Belfrage.
* It's hard to follow the logic in `Belfrage.handle/1` because of the
branching in logic: e.g. if there is a cached response, the request
pipeline and everything that follows is skipped. Also a cached response
is processed differently to e.g. a response that is generated by the
request pipeline.
* It's hard to follow the logic in `Belfrage.handle/1` and
`Cascade.dispatch/1` because at various steps of processing the request
the data that's being passed from function to function can be either a
list with a single `Struct`, a list with multiple `Struct`s or just a
`Struct` and the logic varies accordingly.

I attempted to address the issues above by introducing a `Cascade`
struct, which contains the list of `Struct`s that need to be processed
(`items`) and the `Struct` that contains the response that should be
returned to the client (`result`).

I think it's an improvement because:

* It's a step towards making the cascade functionality a first-class
citizen in the codebase: it makes it clearer that we treat any incoming
request as something that could be passed on to multiple origins.
* It allows to pass a single data structure between functions, reduces
the amount of conditional logic and therefore makes it clearer.
* It allowed to simplify `Belfrage.handle/1`: it now contains the
high-level algorithm for processing an incoming requests and all the
details are in `Belfrage.Cascade` and `Belfrage.Processor`.