CONNPOL-6744 - Add route for Northern Ireland 2017 constituency results (#1077)
[CONNPOL-6744](https://jira.dev.bbc.co.uk/browse/CONNPOL-6744)

Adds a route for the Northern Ireland 2017 Constituency results page on the test environment. The page will be similar to https://www.bbc.co.uk/news/politics/northern-ireland-constituencies/N06000001, but built using our WebCore containers. This is part of the work we are doing for the 2022 Northern Ireland Assembly elections, using the 2017 data to prepare.