Split Authentication.Flagpole into two processes
Belfrage needs to check the status of BBC ID services to determine if
personalisation should be enabled or not. If BBC ID is unavailable or
degraded, personalisation must be turned off.

Before this change the status of BBC ID services was stored in a process
called `Belfrage.Authentication.Flagpole`. That same process also
periodically fetched the status from an API endpoint. The latter
operation could become slow and therefore slow down the code that called
the process to determine if personalisation should be enabled or not,
which effectively means slowing down processing of incoming requests.
Worst case scenario is that a call to determine the status of the
flagpole times out and the request can't be processed at all.

This commit includes the following changes:

* Renames `Belfrage.Authentication.Flagpole` to
`Belfrage.Authentication.BBCID`. I think that the new name is clearer
because 'flagpole' is a term used by BBC ID and one needs to be aware of
what it means to understand it. `BBCID.available?` is hopefully clear
without knowledge of how BBC ID works or its terminology.
* Splits the process into two: `BBCID` will maintain the status and
`BBCID.AvailabilityPoller` will periodically get the status from the
IDCTA config endpoint and update it in `BBCID`. This way if the IDCTA
config endpoint becomes slow or unavailable, it won't affect the code
that checks the status and therefore won't affect processing of incoming
requests.
* Both processes will now be started in tests, but polling is
effectively disabled by setting the interval to an arbitrary high value.
I think this approach is better because there's no difference in what
processes are started between test and prod environments (it's good to
have the test env mirroring the prod env), but there won't be any calls
to the mocked authentication client in the background when tests are
executed, and so no unexpected/confusing test failures.