Track the time it takes to set Struct.private.personalised_request
I initially didn't add tracking for this because I thought that this
operation didn't include anything that could take significant time.
However we now realised that things like e.g. checking the status of BBC
ID can be slow (or even time out) when the process that we're sending a
message to has a lot of messages to processs. We therefore should track
this operation because it's performed for all requests.