Use :telemtry-poller to monitor message queue length
Before this change we had a process called
`Belfrage.Metrics.MailboxMonitor` that periodically retreived message
queue length of a number of processes and reported it to StatsD. There
is a 3rd party library that is capable of doing that which we are
already using: `:telemetry-poller`. It can be configured to periodically
collect info about processes and emit telemetry events. There's also a
library that can transform telemetry events to metrics and send them to
StatsD which we are already using as well: `Telemtry.Metrics`. I thought
that we could replace our own custom process with `:telemetry-poller` +
`Telemetry.Poller`, i.e. effectively replace our own custom logic with
configuration for those 3rd party libraries.

The resulting configuration is still quite a bit of code (not much less
than the logic we had before), which is partly because we would like to
monitor message queue length of loop processes which don't have
registered names, so need to be tracked as custom metrics. However I
still think this is an improvement because we'll have a single way of
doing periodically collecting and reporting metrics (`:telemetry-poller`
+ `Telemetry.Metrics`) instead of multipe ways (our own custom processes
+ `Statix` vs `:telemtry-poller` + `Telemetry.Metrics`).

We used to have tests for `Belfrage.Metrics.MailboxMonitor` and we don't
have tests for the new setup. I think it makes sense because it's
configuration, not logic, and we should not be testing 3rd party
libraries. This is perhaps another advantage of this change: fewer tests
to maintain.

I also removed the test for `Belfrage.Metrics.Supervisor`: it needed to
be updated because the child that it previously started in test env
(`Belfrage.Metrics.MailboxMonitor`) was a worker, but is now a
supervisor (`Belfrage.Metrics.ProcessMessageQueueLength.Supervisor`).
Therefore when `Belfrage.Metrics.Supervisor` is killed in one of the
tests, it was necessary to wait for the child supervisor to terminate,
which wasn't a trivial thing to add. I thought that the test didn't make
much sense to begin with: we don't need to test the functionality of
`Supervisor` (that it restarts its children, etc), so removed the test
instead of trying to fix it.