[RESFRAME-4433] Add max_saturation metric for pools
Pool saturation metrics are currently emitted per pool with the tag
`:pool_name` in `TelemetrySupervisor`. This means that the pool name
is appended to the metric name - as a consequence aggregation cannot
be done based on the metric name in AWS Cloudwatch.

Even if the pool name is moved to the (Cloudwatch) metric dimensions,
for example:

    pool_name: OriginSimulator

AWS Cloudwatch would still not be able to aggregate across pool
dimensions. This is because [1]:

> CloudWatch does not aggregate across dimensions for your custom metrics.

also [2]:

>  You can only retrieve statistics using combinations of dimensions that you specifically published.

This means that the maximum saturation across the pools specified
in the belfrage application must come from a metric emitted from
the said application.

Further aggregation for this metric accross nodes in a given stack
can occur in AWS Cloudwatch.

This commit adds an event called `[:poolboy, :pools]` that contains
aggregate measurements for the HTTP client and AWSExRay pools.
The only measurement at present is the `max_saturation`.

The `track_pool_aggregates/1` function emits the said event.

[1] https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/cloudwatch_concepts.html#dimensions
[2] https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/cloudwatch_concepts.html#dimension-combinations