Configure AWS Agent to collect Statsd metrics
"statsd": {
  "service_address": ":8125",
  "metric_separator": ".",
  "metrics_collection_interval": 1,
  "metrics_aggregation_interval": 60
}

> metrics_collection_interval – How often in seconds that the StatsD plugin runs and collects metrics.
https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Agent-custom-metrics-statsd.html
It made sense to set this to 1 as we want to collect metrics every second for Statsd.

This is not to be confused with the agent level `metric_collection_interval` which I have unset, the current value was
the default anyway. The reasoning behind unsetting this is we want to ensure it is not set to too low of a value as it can cause metrics to
become High-resolution metrics. This seemingly caused gaps in metrics when testing this.
As explained in the docs for Agent `metric_collection_interval`:
> If you set this value below 60 seconds, each metric is collected as a high-resolution metric. For more information about high-resolution metrics, see High-resolution metrics.
https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Agent-Configuration-File-Details.html