Use Personalisation.applicable_request? in ResponseHeaders.Vary
I think dong it this way may be marginally better because:

* The knowledge that personalisation is enabled on bbc.co.uk and not
enabled on bbc.com remains in `Personalisation` module and doesn't leak
into other places.
* The `x-id-oidc-signedin` is only removed from the `Vary` header if the
request is made to a personalised route. This means that we don't need
to set the request host in all tests, which makes it look like the host
is always relevant, when in reality it's only relevant to
personalisation.