Add a test helper to wait for a condition to become true
This can be helpful when testing asynchronous code and there's no other
way to synchronise the result other than polling.

This approach is used in a couple of tests, but the new helper is an
improvement because it limits the polling time. Without the limit it
would just wait for the condition to be matched until the test timeout
is reached (which is currently 2 minutes). The test would then fail
because of the timeout, which could be confusing. This new helper fails
with a human-friendly message instead.