Allow all query params on pres-test routes (#1092)
The pres-test routes are a useful playground for spiking new features and innovation time projects etc. and a number of people have been caught out recently by the param stripping taking place.

The allowlist was put in place in this PR https://github.com/bbc/belfrage/pull/684 by the search team. If it's still required we could consider creating a new route spec for the particular route that needed it but I can't see any references to it the web repo.