[RESFRAME-4433] Remove pool saturation measurement per pool
The pool saturation measurement will not be reported per pool
anymore, just the aggregation (max pool saturation). This will be
used by the scaling policy. If we need to determine which
pool had the maximum saturation for a given interval we can look
at the available workers, overflow workers and the pool size for
each pool.