Disable dials polling in tests
`Belfrage.Dials.Poller` periodically reads the dials file and updates
the dials with the values from that file. We don't want this to happen
in tests because in some tests we update the values of dials and so we
don't want this process to overwrite those changes: it would be very
confusing.

Before this change this was achieved by not starting the `Poller`
process at all in `Belfrage.Dials.Supervisor` in test environment. This
commit replaces that approach with a configurable polling delay, which
allows to effectively disable polling by setting the configuration
parameter to an arbitrary high value in tests. Polling still doesn't
happen in tests, but I think that this approach is better because it
makes all the various environments less different: at least we're
starting the `Poller` process in tests.

I also removed various functions from `Poller` (like `refresh_now`,
`clear`, `state`) that are not actually used anywhere.