Fetch JWK keys from the API in a separate process
Before this change we had a single `Belfrage.Authentication.Jwk` process
that both maintained JWK keys and periodically updated them by calling
an API endpoint. This was a problem, because if fetching from the API
became slow it would slow down the code that needed to get the keys from
the process. JWK keys are used when parsing user session tokens, i.e.
for all requests to personalised routes by signed in users, so this
issue could slow down a significant proportion of traffic.

After this change we have two processes:

* `Belfrage.Authentication.JWK` that maintains the keys in its state (it
still reads the so-called "static" keys from a file on startup).
* `Belfrage.Authentication.JWK.Poller` which periodically makes calls to
the API to get a fresh set of keys.

Another change that this commit includes is that polling of the JWK API
endpoint is now disabled in tests by setting a new `jwk_polling_enabled`
configuration parameter to `false`. It's good to have polling disabled
in tests because it requires the authentication client to be mocked and
if it's not then it can cause unexpected failures in unrelated tests.
Polling was previously disabled too because we just wouldn't start the
JWK process in tests by default. I think that having a configuration
parameter makes it a bit clearer because all environment-specifc
configuration/differences are kept in a single place: config files.

I also replaced the mechanism for loading "static" keys: it used to
happen on compilation and for that reason involved a bit of "magic" and
wasn't easy to test. I replaced it with reading keys on startup of the
`JWK` process which is more straightforward and easier to unit-test.