Make 'on' the default value of the personalisation dial
Personalisation is now always 'on' in production, so making this change
makes sense on its own. However another reason to make it is that our
tests expect the personalisation dial to be 'on' and because polling of
the dials file is now disabled in test, the dials maintain their default
values. Before `Dials.Poller` would read the values of the dials from
`test/support/resources/dials.json` where the value of the
personalisation dial was 'on'.