Only use a dials values file in dev env
Now when polling of the dials file in tests is disabled it can be
confusing to have a file in `test/support/resources` directory that
isn't actually read in tests: one can assume that dials in tests should
have the values from that file when in fact they have the default values
from `cosmos/dials.json`. Therefore I thought that it'd be best to
remove that file from `test/support/resources`.

I think we still want to have a file with dials values in dev
environment where polling is enabled: if there's no file to read
there'll be errors, plus one may want to try to modify that file to
change the value of the dials when testing something locally. I
therefore moved the file with dials values from
`test/support/resources/dials.json` to
`cosmos/dials_values_dev_env.json`. I'm hoping that the name will make
it clear that it's only used in dev env.