Refactor handling credentials for Webcore lambda
Belfrage needs AWS credentials to make calls to Webcore lambda. It gets
them by assuming a AWS IAM role that is allowed to make calls to the
lambda. Belfrage needs to maintain those credentials and periodically
update them.

I started looking into this because we had a single process/module
(`Belfrage.Credentials.Referesh`) that was used to both get the
credentials from AWS, update them and provide them to the code that
calls the lambda. I thought that it could be an issue because if getting
credentials from AWS becomes slow it would slow down the code that used
the credentials to call the lambda. But upon investigation, I realised
that it wasn't actually an issue because the credentials were stored in
ETS, not in the process (the process only periodically updated them).

Nevertheless I thought it was a good opportunity to improve this part of
the system a bit. I thought there were these issues with it:

* Because the module was called `Belfrage.Credentials` it made it look
like a general purpose storage for credentials, when in fact it was only
capable of getting/storing credentials for accessing the Webcore lambda.
In general, naming in `Belfrage.Credentials.*` modules wasn't always
clear.
* Storing the credentials in ETS wasn't really necessary: storing it in
the process state (if it was a dedicated process) would be absolutely
fine. If it was a dedicated process just for maintaining the
credentials, we wouldn't need to worry about it crashing (because it
doesn't do anything) and we don't need access to credentials from
multiple processes, so we don't really need ETS here.
* There weren't any tests for this functionality.

I made the following changes in this PR:

* Added `Belfrage.AWS.Credentials` struct that represents AWS
credentials. I thought that it would be good to take advantage of the
help that the compiler can provide us if we use a struct instead of a
map.
* Moved the code to `Belfrage.Services.Webcore.Credentials` "namespace".
It's only relevant to Webcore, so hopefully this will make it clearer.
* Added a `Belfrage.Services.Webcore.Credentials` process for keeping
credentials in process state (instead of ETS). It gets the credentials
on startup.
* Moved the logic for periodically updating the credentials to
`Credentials.Poller`: it only does that single thing now.
* Added multiple "sources" for credentials: `Credentials.STS`, which is
used normally and `Credentials.Env` which is used in dev and test
environments (these are essentiall credential "strategies" that we had
before.
* Made sure that the `Webcore.Credentials` process is always started in
all environments. The poller isn't started in test or dev environments
which is controlled by a new `webcore_credentials_polling_enabled`
configuration parameter. Before the process that updated the credentials
wasn't started in tests in the supervisor based on the environment. I
thought that adding a configuration parameter for that should make it
clearer because all configuration will be in a single place: config
files.
* Updated `Belfrage.Clients.Lambda` so that it accepts credentials
instead of trying to get them. I think it makes sense because the client
can theoretically be used for multiple things, but the credentials we've
got now are for Webcore only.