Remove test for Authentication.Supervisor
We've noticed that tests for `Belfrage.Authentication.JWK` would
sometimes fail because the `JWK` process would not be running:

```
 1) test get/2 returns static keys on startup (Belfrage.Authentication.JWKTest)
     test/belfrage/authentication/jwk_test.exs:8
     ** (exit) exited in: GenServer.call(Belfrage.Authentication.JWK, {:get, #Function<0.27338317/1 in Belfrage.Authentication.JWK.get/3>}, 5000)
         ** (EXIT) no process: the process is not alive or there's no process currently associated with the given name, possibly because its application isn't started

     stacktrace:
       (elixir 1.11.3) lib/gen_server.ex:1017: GenServer.call/3
       (belfrage 0.2.0) lib/belfrage/authentication/jwk.ex:23: Belfrage.Authentication.JWK.get/3
       test/belfrage/authentication/jwk_test.exs:14: anonymous fn/2 in Belfrage.Authentication.JWKTest."test get/2 returns static keys on startup"/1
```

When I looked into this I realised that we had an async test for
`Authentication.Supervisor` that killed the supervisor process. That
obviously results in restarting the supervisor and all its children, and
in the meantime those children won't be running / available. This means
that if a test that depends on those runs at the same time (any async
test) then it might fail, which is what I think we saw happen to the
`JWK` tests.

I think that we don't really need to test that a supervisor, when killed
is restarted because that's the functionality that OTP provides, so I
just removed the test for `Authentication.Supervisor` instead of trying
to fix it.