Add some docs about metrics
While investigating an issue with our Grafana dashboards I realised that
I don't really understand how our metrics setup works, so I tried to
figure it out. I then thought that it would be good to write down what I
found in case it's helpful in future.

This is far from a complete doc about metrics, but I thought this could
be a starting point which we would then expand. It also mentions what we
could improve (when it comes metric dimensions) and arguably docs are
not the best place for that kind of thing, but I didn't know where else
to put and also that it hopefully helps to clarify the setup.