Read error pages into module attributes on compilation
Before this change we read the error pages from disk every time we
needed to render them. This was inefficient because the pages never
change, so we could just read them once and serve that version.

Reading error pages into module attributes on compilation is probably
this simplest / most efficient way to do it, so this commit implements
that.

Because we're now reading the error pages in all envs, I moved them into
`priv/static/default_error_pages`. Keeping them in `test/support`
directory made it look like they are only used in tests.