Refactor Belfrage.Service.WebcoreTest
Removed details that don't make a difference from the test to make it
less wordy and hopefully clearer.

The test tests `Service.Webcore` itself and the logic in
`Service.Webcore.Response` because I'm planning to remote the latter in
a following commit.

It turns out `:b64fast` doesn't actually raise `ArgumentError` when the
input is not base64 encoded: it just attempts to decoded it in any case.
I don't think this is a major problem, but added a test with a comment
that makes it clear. The relevant test in `Service.Webcore.ResponseTest`
is invalid: the 500 response is returned not because of the base64
encoding, but because of the invalid format of the response from Lambda.