Refactor Services.Webcore
Moved the logic from `Services.Webcore.Response` to `Services.Webcore`
because I'm planning to add more response-related metrics that would
depend on the request and I didn't want to pass the request to
`Services.Webcore.Response`.

There wasn't that much logic in `Webcore.Response` anyway: it relied on
multiple function clauses quite a bit, and using control flow structures
instead makes it more compact, so it doesn't make `Services.Webcore` too
big. It's also fairly easy to test that logic in the test for
`Services.Webcore`.